requirements.txt
torch>=1.10.0
torchvision>=0.11.0
fastai>=2.7.0
lightly>=1.4.0
datasets>=2.0.0
transformers>=4.20.0
diffusers>=0.14.0
hydra-core>=1.3.0
omegaconf>=2.3.0
wandb>=0.13.0
scikit-learn>=1.0.0
umap-learn>=0.5.0
matplotlib>=3.5.0
seaborn>=0.12.0
tqdm>=4.64.0
pillow>=9.0.0
numpy>=1.20.0
accelerate>=0.16.0
huggingface_hub>=0.14.0
